end game and randomized spawning

+++ SummativeGame/src/screen/DisplayManager.java
@@ -67,7 +67,7 @@ public class DisplayManager {
         endTime = System.currentTimeMillis();
         if(won == null){
             won = new GameWon(this, ScreenVariables.getGameWonName(), 
-                    ScreenVariables.getGameWonWidth(), ScreenVariables.getGameWonHeight(), ((endTime-startTime)/1000));
+                    ScreenVariables.getGameWonWidth(), ScreenVariables.getGameWonHeight(), (double)((endTime-startTime)/1000));
         }else{
             won.frameVisibility(true);
         }
